+++ internal/provider/data_source_meraki_wireless_rf_profile.go
@@ -59,7 +59,7 @@ func (d *WirelessRFProfileDataSource) Metadata(_ context.Context, req datasource
 func (d *WirelessRFProfileDataSource) Schema(ctx context.Context, req datasource.SchemaRequest, resp *datasource.SchemaResponse) {
 	resp.Schema = schema.Schema{
 		// This description is used by the documentation generator and the language server.
-		MarkdownDescription: helpers.NewAttributeDescription("This data source can read the `Wireless RF profile` configuration.").String,
+		MarkdownDescription: helpers.NewAttributeDescription("This data source can read the `Wireless RF Profile` configuration.").String,
 
 		Attributes: map[string]schema.Attribute{
 			"id": schema.StringAttribute{