+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Add `meraki_organization_adaptive_policies` resource
 - Add `meraki_organization_adaptive_policy_groups` resource
 - Add `meraki_wireless_ssid_identity_psks` resource
+- Add `meraki_wireless_rf_profiles` resource
 
 ## 1.3.0